% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Generics.R, R/rQivivo-Methods.R
\name{absence}
\alias{absence}
\title{ACTIVE or DELETE the absence}
\usage{
absence(object, request, body)
}
\arguments{
\item{object}{an rQivivo S4 object.}

\item{request}{character:
\describe{
   \item{"post":}{sets the thermostat into absence mode (i.e. the frost temperature);}
   \item{"del":}{removes the absence mode.}
}}

\item{body}{list: only if \code{request = "post"}:
\describe{
   \item{start_date POSIXt:}{the starting date of the absence;}
   \item{end_date POSIXt:}{the ending date of the absence.}
}}
}
\description{
The function changes the thermostat into the frost temperature or cancels the absence mode.
}
\author{
Theo DEVAUCOUP \email{theo.devaucoup@pm.me}
}
