% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Generics.R, R/rQivivo-Methods.R
\name{arrival}
\alias{arrival}
\title{ACTIVE or DELETE the arrival}
\usage{
arrival(object, request, duration)
}
\arguments{
\item{object}{an rQivivo S4 object.}

\item{request}{character:
\describe{
   \item{"post":}{sets the thermostat into arrival mode (i.e. to start heating);}
   \item{"del":}{removes the arrival mode.}
}}

\item{duration}{numeric: duration before the arrival, to start heating (in minutes).}
}
\description{
The function starts or stops heating the house earlier than the programmation was anticipating.
}
\author{
Theo DEVAUCOUP \email{theo.devaucoup@pm.me}
}
