% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Generics.R, R/rQivivo-Methods.R
\name{getThermostat}
\alias{getThermostat}
\title{GET some information about the thermostat}
\usage{
getThermostat(object, type)
}
\arguments{
\item{object}{an rQivivo S4 object.}

\item{type}{character: type of \code{\link[httr]{GET}} request:
\describe{
   \item{"info":}{gives some information about the device;}
   \item{"temperature":}{gives the temperature;}
   \item{"humidity":}{gives the humidity;}
   \item{"presence":}{gives the last presence.}
}}
}
\value{
The function returns the content of the get request.
}
\description{
The function provides get resquest about the thermostat.
}
\author{
Theo DEVAUCOUP \email{theo.devaucoup@pm.me}
}
