% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Generics.R, R/rQivivo-Methods.R
\name{refresh}
\alias{refresh}
\title{Refresh the OAuth connection}
\usage{
refresh(object)
}
\arguments{
\item{object}{an rQivivo S4 object.}
}
\description{
The function refresh the current authorization with the \code{refresh_token}.
}
\author{
Theo DEVAUCOUP \email{theo.devaucoup@pm.me}
}
