% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Class.R
\name{rQivivo-class}
\alias{rQivivo-class}
\title{rQivivo class}
\description{
rQivivo class with OAuth2.0 using httr for API connection.
See \href{https://documenter.getpostman.com/view/1147709/qivivo-api/2MsDNL#d64f1f8b-f8e9-6470-50d5-980393886046}{Qivivo API}.
}
\section{Slots}{

\describe{
\item{\code{token}}{Token2.0: return of \code{\link[httr]{oauth2.0_token}} function.}

\item{\code{UUID}}{character: UUID of the thermostat.}
}}

