% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Class.R
\name{rQivivo}
\alias{rQivivo}
\title{rQivivo object constructor}
\usage{
rQivivo(appName, clientID, secretID, redirectURI)
}
\arguments{
\item{appName}{character: (optional) name of the private API.}

\item{clientID}{character: client ID of the private API.}

\item{secretID}{character: secret ID of the private API.}

\item{redirectURI}{character: redirected URI of the private API. Use locate IP for RStudio.}
}
\value{
The function returns an rQivivo object.
}
\description{
The function is the constructor of rQivivo class. See \url{https://account.qivivo.com/}.
}
