% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rQivivo-Class.R
\name{absence}
\alias{absence}
\title{POST or GET the absence}
\usage{
absence(object, request, body)
}
\arguments{
\item{request}{character:
\itemize{
   \item post: sets the thermostat into absence mode (i.e. the frost temperature);
   \item del: removes the absence mode.
}}

\item{body}{list: only if \code{request = "post"}:
\itemize{
   \item start_date POSIXt: the starting date of the absence;
   \item end_date POSIXt: the ending date of the absence.
}}
}
\description{
The function changes the thermostat into the frost temperature or cancels the absence mode.
}
